♻️ Moved sortIds() and applied a MohistMC's version list
